requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
Pillow>=10.2.0
pytest==7.4.3
pytest-cov==4.1.0
openai>=1.40.0,<2.0.0
tiktoken>=0.7.0

